refactor: Admin Space DTO를 dto/req, dto/res 폴더로 분리 및 불필요 파일 정리 (#59)
